暂存

+++ src/xens/gui/thread/DecryptThread.java
@@ -18,7 +18,8 @@ public class DecryptThread extends Thread {
     JComboBox<String> decryptMethod;//解密方法
     JTextArea consoleArea;//输出域
     String priKeyPath;//密钥文件地址
-    public DecryptThread(JFrame jf,JButton btnEncrypt,JButton btnDecrypt, JTextField decryptFilePath,JTextField decryptKey,JComboBox<String> decryptMethod,JTextArea consoleArea){
+
+    public DecryptThread(JFrame jf, JButton btnEncrypt, JButton btnDecrypt, JTextField decryptFilePath, JTextField decryptKey, JComboBox<String> decryptMethod, JTextArea consoleArea) {
         this.jf = jf;
         this.btnEncrypt = btnEncrypt;
         this.btnDecrypt = btnDecrypt;
@@ -29,7 +30,8 @@ public class DecryptThread extends Thread {
         this.setName("Encrypt_Thread");
 
     }
-//    public static readPriKey(JTextField decryptKey){
+
+    //    public static readPriKey(JTextField decryptKey){
 //        String path = decryptKey.getText();
 //        //创建文件类
 //        File file = new File(path);
@@ -45,7 +47,7 @@ public class DecryptThread extends Thread {
 //        }
 //    }
     //创建线程类
-    public void run(){
+    public void run() {
         readPriKey();
         //获取文件加密路径
         String EncryptPath = decryptFilePath.getText();
@@ -53,20 +55,24 @@ public class DecryptThread extends Thread {
         String[] pathList = EncryptPath.split(",");
         //获取数组长度
         int listLen = pathList.length;
+        //成功个数
+        int successNum = 0;
+        //失败个数
+        int failNum = 0;
         Date start = new Date();
         for (int i = 0; i < listLen; i++) {
             final int index = i;
             final String path = pathList[index].trim();
             //创建文件类
             File file = new File(path);
-            if(!file.exists() || !file.isFile()) {
-                JOptionPane.showMessageDialog(jf, "文件路径错误!", "错误",JOptionPane.WARNING_MESSAGE);
+            if (!file.exists() || !file.isFile()) {
+                JOptionPane.showMessageDialog(jf, "文件路径错误!", "错误", JOptionPane.WARNING_MESSAGE);
                 return;
             }
             //密钥为空时报错
-            if (decryptKey.equals("")){
-                consoleArea.append("密钥错误:"+path);
-                JOptionPane.showMessageDialog(jf, "请检查密钥!", "错误",JOptionPane.WARNING_MESSAGE);
+            if (decryptKey.equals("")) {
+                consoleArea.append("密钥错误:" + path);
+                JOptionPane.showMessageDialog(jf, "请检查密钥!", "错误", JOptionPane.WARNING_MESSAGE);
                 return;
             }
             //创建文件加密实例
@@ -75,45 +81,53 @@ public class DecryptThread extends Thread {
             btnEncrypt.setEnabled(false);
             btnDecrypt.setEnabled(false);
             int resIndex = 0;
-            try{
-                resIndex = fileEncrypter.decrypt(file,path, decryptMethod.getSelectedIndex(),decryptKey);
-            }catch (Exception e){
-                JOptionPane.showMessageDialog(jf,"文件加密出现错误...","加密错误",JOptionPane.WARNING_MESSAGE);
+            try {
+                resIndex = fileEncrypter.decrypt(file, path, decryptMethod.getSelectedIndex(), decryptKey);
+                if (resIndex == 0) {
+                    failNum++;
+                } else if (resIndex == 1) {
+                    successNum++;
+                }
+            } catch (Exception e) {
+                JOptionPane.showMessageDialog(jf, "文件加密出现错误...", "加密错误", JOptionPane.WARNING_MESSAGE);
             }
             //只有一个
-            if (listLen==1){
-                if (resIndex == 1){//解密成功
+            if (listLen == 1) {
+                if (resIndex == 1) {//解密成功
                     consoleArea.append("文件解密完成!");
-                    File file1 = new File(priKeyPath);
-                    if (file1.exists()&&file1.isFile()){
-                        file1.delete();
+                    if(priKeyPath!=null){
+                        File file1 = new File(priKeyPath);
+                        if (file1.exists() && file1.isFile()) {
+                            file1.delete();
+                        }
                     }
+
                     //显示面板
-                    JOptionPane.showMessageDialog(jf,"文件解密成功!","解密成功",JOptionPane.PLAIN_MESSAGE);
-                }else {
+                    JOptionPane.showMessageDialog(jf, "文件解密成功!", "解密成功", JOptionPane.PLAIN_MESSAGE);
+                } else {
                     consoleArea.append("--------------------------------------------------------------------------");
                     consoleArea.append("文件解密失败!!!");
-                    JOptionPane.showMessageDialog(jf,"文件解密出现错误,请检查加密方式和密码...","解密错误",JOptionPane.WARNING_MESSAGE);
+                    JOptionPane.showMessageDialog(jf, "文件解密出现错误,请检查加密方式和密码...", "解密错误", JOptionPane.WARNING_MESSAGE);
                 }
-            }else {//有多个文件
-                if (resIndex == 1){//解密成功
+            } else {//有多个文件
+                if (resIndex == 1) {//解密成功
                     consoleArea.append("文件解密完成!");
-                    if (index==listLen-1){
-                          JOptionPane.showMessageDialog(jf,"文件解密成功!","解密成功",JOptionPane.PLAIN_MESSAGE);
+                    if (index == listLen - 1) {
+                        JOptionPane.showMessageDialog(jf, "文件解密成功!", "解密成功", JOptionPane.PLAIN_MESSAGE);
                     }
-                }else {
+                } else {
                     consoleArea.append("--------------------------------------------------------------------------");
                     consoleArea.append("文件解密失败!!!");
-                    if (index==listLen-1){
-                        JOptionPane.showMessageDialog(jf,"文件解密出现错误,请检查加密方式和密码...","解密错误",JOptionPane.WARNING_MESSAGE);
+                    if (index == listLen - 1) {
+                        JOptionPane.showMessageDialog(jf, "文件解密出现错误,请检查加密方式和密码...", "解密错误", JOptionPane.WARNING_MESSAGE);
                     }
                 }
             }
-
-
         }
         Date end = new Date();
-        consoleArea.append("\r\n解密总耗时:"+TimeFormat(end.getTime() - start.getTime()));
+        consoleArea.append("\r\n总共加密" + listLen + "个文件,其中成功个数为:" + successNum + ",失败个数为:" + failNum + "。");
+
+        consoleArea.append("\r\n解密总" + TimeFormat(end.getTime() - start.getTime()));
 
         //设置按钮可用
         btnEncrypt.setEnabled(true);
@@ -124,14 +138,14 @@ public class DecryptThread extends Thread {
         String path = decryptKey;
         //创建文件类
         File file = new File(path);
-        if(file.exists() && file.isFile()) {
-            try{
+        if (file.exists() && file.isFile()) {
+            try {
                 FileInputStream is = new FileInputStream(path);
                 byte[] priKey = new byte[1024];
                 is.read(priKey);
                 this.decryptKey = new String(priKey);
-                priKeyPath=path;
-            }catch (Exception e){
+                priKeyPath = path;
+            } catch (Exception e) {
                 System.out.println(e);
             }
         }
